0.5.4: support nix-2.9.0
They really did a number on this one. Positions are now interned
(genuinely a great idea! just lots of pain for me), and I want to
support the old version too. So my approach was treating it as opaque
except in compat.h where we hide the skeletons. Some signature changes
too.

Eventually my entire program will be compat::.